[compute anime] save point
rewrite sizing function and propagation and smoothing

+++ Computer_Animation/final/QuadtreeLiquidSimulator/include/Base.h
@@ -8,7 +8,7 @@ struct Particle {
 };
 
 struct Line {
-    int x1, y1, x2, y2;
+    float x1, y1, x2, y2;
 };
 
 class BaseSimulator {
